Flags markdown table is generated from json
