+++ .github/workflows/linux-eic-shell.yml
@@ -136,7 +136,7 @@ jobs:
           export DETECTOR_CONFIG=${DETECTOR}_${{ matrix.detector_config }}
           export JANA_PLUGIN_PATH=$PWD/lib/EICrecon/plugins${JANA_PLUGIN_PATH:+:${JANA_PLUGIN_PATH}}
           chmod a+x bin/*
-          $PWD/bin/run_eicrecon_reco_flags.py sim_pi_1GeV_20GeV_${{ matrix.detector_config }}.edm4hep.root rec_pi_1GeV_20GeV_${{ matrix.detector_config }} -Pdump_flags:json=${{ matrix.detector_config }}_reco_flags_flags.json
+          $PWD/bin/run_eicrecon_reco_flags.py sim_pi_1GeV_20GeV_${{ matrix.detector_config }}.edm4hep.root rec_pi_1GeV_20GeV_${{ matrix.detector_config }} -Pdump_flags:json=${{ matrix.detector_config }}_all_good_flags.json
     - uses: actions/upload-artifact@v3
       with:
         name: run_eicrecon_reco_flags_rec_pi_1GeV_20GeV_${{ matrix.detector_config }}.edm4eic.root
@@ -144,5 +144,22 @@ jobs:
         if-no-files-found: error
     - uses: actions/upload-artifact@v3
       with:
-        path: ${{ matrix.detector_config }}_reco_flags_flags.json
+        path: ${{ matrix.detector_config }}_all_good_flags.json
         if-no-files-found: error
+
+generate-docs:
+  runs-on: ubuntu-latest
+  needs:
+    - run_eicrecon_reco_flags-gcc
+    - eicrecon-gcc
+  steps:
+    - uses: actions/download-artifact@v3
+      with:
+        name: arches_all_good_flags.json
+    - uses: cvmfs-contrib/github-action-cvmfs@v3
+    - uses: eic/run-cvmfs-osg-eic-shell@main
+      with:
+        platform-release: "jug_xl:nightly"
+        setup: /opt/detector/setup.sh
+        run: |
+          ls -ll
